Stop confusing a removed label from a grah with a new one.
Actually related to pr_grapher_wh_nagios and UI.

Previously, when we were removing a label from a graph by editing it in
the UI, if no other graphs were using this label, the UI creating a new
graph to stick the removed label in it. There was confusion between new
labels freshly created from a service and never seen in the graphs, and
old labels removed from graphs.
A new graph is created for new labels based on its existance in table
"pr_grapher.graph_wh_nagios". This patch makes sure any known labels
always stays in there, even if removed from all graphes, by setting
graph_id to NULL for the very last occurence of the label.
The new PL/PgSQL function pr_grapher.create_graph_for_wh_nagios(...)
takes care of this logic and is the first step to the full and secured
API of this module.