Refactoring of "Change graph/service" lists in graph pages
Integrate the lists in the same navbar than period selection to save
some more time.
Change the behavior of the list in the showservice page. The list was
showing all the graph and was jumping on a sepcific graph only. Now, it
shows only the services, and jump to another showservice page.