Removes links to graphs from the server page
On each server page, we used to show services and graphs for each
services, which makes the page and workflow difficult to understand.
This patch removes graphs links from the server page. These links will
be available from the services pages next to their corresponding graph
in a futur patch.