pg: Handle warehouse deletion.
Previously, if a warehouse which registedered some API function was dropped,
public.grant_appli() would fail as the corresponding functions would not exist
anymore.  To fix this issue, an event trigger on DROP EXTENSION is created
which tries to catch all OPM warehouses deletion.  We consider that it's an
OPM extension if the dropped funtions' schema starts with 'wh_'.  A NOTICE is
raised with the schema name if some functions where deleted from public.api.