ui: Add ability to specify interval in the URL.
It's the first step to have a complete feature. For now, the graph link button
won't use the specified interval, and tags will remove them.

The interval must be the las parameter in graphs url, in the format :
timestamp_string1 | timestmap_string2
The spaces around | are optionnal and will be removed, whatever the number of
spaces. The timestamp_string must be parsable by moment.js without format, see
http://momentjs.com/docs/#/parsing/string/ for complete list.