Handles services without any metrics.
It can be interesting to display the services and related status, even if this
service don't have any data to graph. Some recent version of nagios can generate
files with empty perfdata, thus allowing OPM to report a status change. In case
of a service which never generate perfdata (check_configuration for instance),
this would result in a link to an always empty page.